requirements.txt
# Core dependencies
streamlit
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0

# News and data
requests>=2.31.0
python-dotenv>=1.0.0

# Visualization
graphviz>=0.20.1
pillow>=10.0.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.2

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6